Implement some magic for macros and environment variables
You can now use #macro.<anything> and #envvar.<anything> on Packer::Config instances and it will magically turn the <anything> bits in to proper packer.io macro or environment variable strings. This is is a much cleaner interface than a single method that takes arguments. The updated README.md file shows the changes in practice.