Finalize support for Packer Communicators
Packer 0.8.5 introduced the concept of Communicators, which are called
out as Templates in the documentation, but actually exist and are
configured as first-class attributes on Builders. Therefore, they've
been implemented as Builder attributes.

Thanks to @diamond29 who got this ball rolling. This just cleans up the
implementiation.

This commit also adds support for the use of LoweredExpectations to
enforce a minimum version of Packer is present on the machine when you
try to run validation on your configuration. Right now the minimum
version is set at 0.8.5 so Communicators support is assured.