Update for latest rubocop, cleanup TODO rubocop rules
Fixes #22

This now works with the latest Rubocop release (0.42.0). I also did a
bunch of cleanup so the TODO list of things of that weren't passing
Rubocop checks is much, much smaller now. There are a handful of TODOs
left that require touching a lot of files. I'll get to those later.